% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dba_gc_ex003}
\alias{dba_gc_ex003}
\title{DBA Group Composition Data Example 003}
\format{
## `dba_gc_ex003`
A matrix with 4 rows and 4 columns

* id: the student id of each students, simply the integers 1 to 4.
* self_groups: The self-formed groups
* year, major: demographics used in computing dissimilarities
}
\source{
This dataset was constructed by hand.
}
\usage{
dba_gc_ex003
}
\description{
An example dataset to use with the diversity-based assignment model. It is
used to demonstrate the use of a custom dissimilarity matrix.
}
\keyword{datasets}
