% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouper-package.R
\docType{package}
\name{grouper-package}
\alias{grouper}
\alias{grouper-package}
\title{grouper: A Shiny App to Group Students in Turns and Rooms for
    Online Exams}
\description{
For Zoom examination it is useful to divided
    students into breakout rooms. When students are in a large number it
    could be necessary to consider multiple turns to complete an
    examination. {grouper} help you in grouping students in turns and in
    breakout room within each turn.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.github.com/CorradoLanera/grouper}
  \item Report bugs at \url{https://www.github.com/CorradoLanera/grouper/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@unipd.it}

}
\keyword{internal}
