% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multirole_prefmat_ex001}
\alias{multirole_prefmat_ex001}
\title{Multi-role Preference Matrix Example 001}
\format{
## `multirole_prefmat_ex001`
A matrix with 4 rows and 4 columns.

Rows correspond to individuals in `multirole_students_ex001`, and columns
correspond to rows of `multirole_demand_ex001`.

Preference scores are encoded as 3 (first choice), 2 (second choice), and 1
(third choice). Unranked courses are encoded as -99.
}
\source{
This dataset was constructed by hand.
}
\usage{
multirole_prefmat_ex001
}
\description{
An example preference matrix for the multi-role workload allocation model.
It can be used for either TA or GR preferences.
}
\keyword{datasets}
