% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phd_demand_ex001}
\alias{phd_demand_ex001}
\title{PhD Demand Matrix Example 001}
\format{
\subsection{\code{phd_demand_ex001}}{

A matrix with 4 rows and 2 columns.

Columns are in the order \code{TA}, \code{GR}. Row names store the course codes.
}
}
\source{
This dataset was constructed by hand.
}
\usage{
phd_demand_ex001
}
\description{
An example demand matrix to use with the PhD workload allocation model.
}
\keyword{datasets}
