% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{prepare_diversity_model}
\alias{prepare_diversity_model}
\title{Prepare the diversity-based assignment model}
\usage{
prepare_diversity_model(df_list, yaml_list, w1 = 0.5, w2 = 0.5)
}
\arguments{
\item{df_list}{The output list from [extract_student_info()] for
`assignment = "diversity"`.}

\item{yaml_list}{The output list from [extract_params_yaml()] for
`assignment = "diversity"`.}

\item{w1, w2}{Numeric values between 0 and 1. Should sum to 1. These weights
correspond to the importance given to the diversity- and skill-based
portions in the objective function.}
}
\value{
An ompr model.
}
\description{
Prepare the diversity-based assignment model
}
