% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{prepare_preference_model}
\alias{prepare_preference_model}
\title{Prepare the preference-based assignment model}
\usage{
prepare_preference_model(df_list, yaml_list)
}
\arguments{
\item{df_list}{The output list from [extract_student_info()] for
`assignment = "preference"`.}

\item{yaml_list}{The output list from [extract_params_yaml()] for
`assignment = "preference"`.}
}
\value{
An ompr model.
}
\description{
Prepare the preference-based assignment model
}
