% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dba_gc_ex004}
\alias{dba_gc_ex004}
\title{DBA Group Composition Data Example 004}
\format{
\subsection{\code{dba_gc_ex004}}{

A matrix with 5 rows and 4 columns
\itemize{
\item id: the student id of each students, simply the integers 1 to 4.
\item self_groups: The self-formed groups
\item python: Python skill level - 1 is lowest, 3 is highest.
}
}
}
\source{
This dataset was constructed by hand.
}
\usage{
dba_gc_ex004
}
\description{
An example dataset to use with the diversity-based assignment model. It is
used to demonstrate the use of a vectors to indicate individual group size
constraints for specific topics.
}
\keyword{datasets}
