% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fns.R
\name{extract_student_info}
\alias{extract_student_info}
\title{Extract student information}
\usage{
extract_student_info(dframe, demographic_cols, skills, self_formed_groups)
}
\arguments{
\item{dframe}{A dataframe with one row for each student. The columns could
possibly contain demographic variables, an overall skill measure, and
a column indicating self-formed groups. There must be a column named
id, for use by other functions in this package.}

\item{demographic_cols}{A set of integers indicating the columns corresponding
to demographic information, e.g. major, year of study, gender, etc.}

\item{skills}{A numeric measure of overall skill level (higher means more
skilled).}

\item{self_formed_groups}{An integer column that identifies the self-formed
groups, submitted by students.}
}
\value{
A list containing:

  N - number of students
  G - number of self-formed groups
  m - a student x groups matrix, indicating which student is in which
      group
  d - dissimilarity matrix
  s - skills vector for each individual student
}
\description{
Converts a dataframe with information on students to a list of parameters. This
list forms one half of the inputs to prepare_model(). The other half comes from
extract_params_yaml.
}
\details{
The demographic variables are converted into an NxN dissimilarity matrix.
}
