% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pba_gc_ex002}
\alias{pba_gc_ex002}
\title{PBA Group Composition Data Example 002}
\format{
\subsection{\code{pba_gc_ex002}}{

A data frame with 8 rows and 2 columns.
\itemize{
\item id: the student id of each students, simply the integers 1 to 8.
\item grouping: the self-formed groups submitted by each student. In this case,
each self-formed group is of size 2.
}
}
}
\source{
This dataset was constructed by hand.
}
\usage{
pba_gc_ex002
}
\description{
An example dataset to use with the preference-based assignment model.
}
\keyword{datasets}
