% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phd_prefmat_ex001}
\alias{phd_prefmat_ex001}
\title{PhD Preference Matrix Example 001}
\format{
\subsection{\code{phd_prefmat_ex001}}{

A matrix with 4 rows and 4 columns.

Rows correspond to students in \code{phd_students_ex001}, and columns correspond
to rows of \code{phd_demand_ex001}.

Preference scores are encoded as 3 (first choice), 2 (second choice),
and 1 (third choice). Unranked courses are encoded as -99.
}
}
\source{
This dataset was constructed by hand.
}
\usage{
phd_prefmat_ex001
}
\description{
An example preference matrix to use with the PhD workload allocation model.
}
\keyword{datasets}
