% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{prepare_model}
\alias{prepare_model}
\title{Initialise model 1}
\usage{
prepare_model(df_list, yaml_list, w1 = 0.5, w2 = 0.5)
}
\arguments{
\item{df_list}{The output list from extract_student_info().}

\item{yaml_list}{The output list from extract_params_yaml().}

\item{w1, w2}{Numeric values between 0 and 1. Should sum to 1.}
}
\value{
An ompr model.
}
\description{
Initialise model 1
}
